requirements.txt
instabot==0.117.0
pillow==7.1.0
python-dotenv==0.12.0
requests==2.23.0
urllib3==1.25.8